Idea2/notification microservice/requirements.txt
Flask==3.1.0
twilio==9.5.0
flask-cors==5.0.0
python-dotenv==1.0.0
pika